added C8, changed C7, C6 is broken
C8 is now the camera+twitter example and won't be included in the
hardware guide. C7 just uses the camera and button, I think this is a
little more stable for the manual. C6 isn't compiling and is stopping on
the flieIO stuff. I found v2.0a4 compiles C6 fine but not the newest
v2.0b7, so some something has changed in either processing or android.